Refactor decorator to different files
