+++ decorator.rb
@@ -12,16 +12,3 @@ class Decorator < Nameable
     @nameable.correct_name
   end
 end
-
-class CapitalizeDecorator < Decorator
-  def correct_name
-    @nameable.correct_name.capitalize
-  end
-end
-
-class TrimmerDecorator < Decorator
-  def correct_name
-    @nameable.correct_name unless @nameable.correct_name.length <= 10
-    @nameable.correct_name[0..9]
-  end
-end
