+++ person.rb
@@ -1,5 +1,4 @@
 require_relative './nameable'
-require_relative './decorator'
 
 class Person < Nameable
   attr_reader :id
@@ -29,13 +28,3 @@ class Person < Nameable
     @name
   end
 end
-
-# Testing Decorator
-person = Person.new(22, 'maximilianus')
-puts person.correct_name # maximilianus
-
-capitalized_person = CapitalizeDecorator.new(person)
-puts capitalized_person.correct_name # Maximilianus
-
-capitalized_trimmed_person = TrimmerDecorator.new(capitalized_person)
-puts capitalized_trimmed_person.correct_name # Maximilian
